APP.tsx、Login.tsx、Game.tsxの追加・修正＆それに伴いJava側の修正
