+++ backend/src/main/java/com/example/generator/PasswordGenerator.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.generator;
 
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
